Service: Refresh architecture diagram w hyperlinks
The architecture diagram now follows a similar style as we use in other
places, e.g. koji-osbuild.
Furthermore, each component of our service now holds a direct link to
the corresponding GitHub repository. For simplicity, I have linked the
'worker fleet' to the osbuild repository.
Also, the legend that was shown below the diagram explaining the
authentication mechanisms between the pieces of the service has been
omitted.

Please note that the hyperlink rendering only works when including the
svg file in markdown via {{#include }}. Furthermore, the svg's source
cannot contain empty newlines or routines that might be interpreted as
markdown. Otherwise it will fail to render correctly.