blueprint: list image types that don't support fs customizations
Add a note for the image types that don't support filesystem
customizations.  This is a common point of confusion among users,
especially for the Anaconda-based ISO image types.

Separate the image types where fs customizations don't make sense, to
suggest that the others might get the support in the future.